Separate logging functions & address translation functions into their  respective own static libraries.
