+++ C/include/addresstranslation.h
@@ -7,10 +7,7 @@
 #ifndef __ADDRESS_TRANSLATION_H
 #define __ADDRESS_TRANSLATION_H
 
-#include <inttypes.h>
 #include <stdint.h>
-#include <stdio.h>
-#include <unistd.h>
 
 uintptr_t virtual_to_physical_address(uintptr_t virt_addr);
